packager: Parse and compile scripts before packaging
The package contains compiled bytecode in JSON format
instead of source files. The benefit is parser/compiler
errors are detected during compilation, and improved
performance.

Additionally JSON files are more difficult to parse for
humans, therefore the script is obfuscated for naked eye.
This should not be relied upon as special tools can be
implemented to reverse bytecode to a source code.